move to generic "parsers" library with entry parser being one element.
