+++ src/lib.rs
@@ -1,9 +1,9 @@
 pub mod cli_processor;
 pub mod config;
 pub mod db;
-pub mod entry;
 pub mod filenames;
 pub mod oauth;
+pub mod parsers;
 pub mod record;
 pub mod time;
 pub mod ui;
